+++ event-store-util/src/test/java/uk/gov/justice/services/eventsourcing/util/jee/timer/TimerConfigFactoryTest.java
@@ -1,4 +1,4 @@
-package uk.gov.justice.services.eventsourcing.timer;
+package uk.gov.justice.services.eventsourcing.util.jee.timer;
 
 import static org.hamcrest.CoreMatchers.instanceOf;
 import static org.hamcrest.CoreMatchers.is;